Moved Memory.cs to root
