+++ README.md
@@ -1,11 +1,15 @@
 A fork of https://github.com/fholm/UnsafeCollections
 
+**This fork is still a WIP**
+
+This fork contains many bug fixes/improvements and an easier to use API.
+
 Project is targeted as a .Net 2.0 Standard library. Usable in Unity via dll.
 
 ## Build
 Use Preprocessor directive UNITY to build the project using the Unity memory allocators instead of the .Net ones.
 
-The library is usable in both .Net as well as Unity.
+The library is usable in both .Net and Unity.
 
 
 ### ToDo
