+++ UnsafeCollections/Collections/Unsafe/UnsafeBitSet.cs
@@ -27,8 +27,6 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using UnsafeCollections;
-using UnsafeCollections.Unsafe;
 #if UNITY
 using Unity.Collections.LowLevel.Unsafe;
 #endif
